Allegro Setup
Setup allegro 5
Not the way i liked but the way it works:
Allegro 5 is setup as nuget package in package folder in Solution root dir. 
The way i want is to setup allegro 5 in Engine/vendor

but i cannot figureout how to make premake work with that. 

Maybe someday compile and link allegro5 from source